CRB-4254: Python 3 compatible job-runner

+++ Cerebrum/modules/job_runner/socket_ipc.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-
-# Copyright 2018-2021 University of Oslo, Norway
+#
+# Copyright 2018-2024 University of Oslo, Norway
 #
 # This file is part of Cerebrum.
 #
@@ -17,22 +17,28 @@
 # You should have received a copy of the GNU General Public License
 # along with Cerebrum; if not, write to the Free Software Foundation,
 # Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307, USA.
-""" Job Runner socket protocol. """
-from __future__ import print_function
-
-import json
+"""
+Job Runner socket server, client, and protocol.
+"""
+from __future__ import (
+    absolute_import,
+    division,
+    print_function,
+    unicode_literals,
+)
+import contextlib
 import logging
 import os
 import signal
 import socket
 import threading
 import time
-from contextlib import closing
 
 from six import text_type
 
 import cereconf
 
+from Cerebrum.utils import json
 from Cerebrum.utils.date import to_seconds
 from .times import fmt_asc, fmt_time
 from .health import get_health_report
@@ -51,7 +57,7 @@ def signal_timeout(signal, frame):
 
 
 class SocketConnection(object):
-    """ Simple socket reader/writer with buffer and size limit. """
+    """ Simple socket reader/writer with buffer and size limit.  """
 
     eof = b'\0'
     buffer_size = 1024 * 2
@@ -62,6 +68,8 @@ class SocketConnection(object):
 
     def send(self, data):
         """ Send a null-terminated bytestring. """
+        if not isinstance(data, bytes):
+            raise ValueError("Expected bytes, got " + repr(type(data)))
         if self.eof in data:
             raise ValueError("payload cannot contain eof")
         data = data + self.eof
@@ -91,6 +99,29 @@ class SocketConnection(object):
             if chunk.endswith(self.eof):
                 break
         return b''.join(chunks)
+
+
+@contextlib.contextmanager
+def jr_connection(path=None, timeout=None):
+    """
+    Get a SocketConnection to job-runner.
+
+    :param str path:
+        Path to the job-runner socket (defaults to cereconf.JOB_RUNNER_SOCKET)
+
+    :param float timeout:
+        Timeout for the socket connection (defaults to no timeout)
+
+    :returns:
+        A SocketConnection context
+    """
+    path = path or cereconf.JOB_RUNNER_SOCKET
+    timeout = float(timeout) if timeout else None
+    with contextlib.closing(socket.socket(socket.AF_UNIX)) as sock:
+        sock.connect(path)
+        if timeout:
+            sock.settimeout(timeout)
+        yield SocketConnection(sock)
 
 
 class Commands(object):
@@ -143,12 +174,24 @@ class Commands(object):
 
 
 class SocketProtocol(object):
-
+    """
+    The job runner socket protocol.
+    """
     commands = Commands()
 
     def __init__(self, connection, job_runner):
-        self.job_runner = job_runner
         self.connection = connection
+        self.job_runner = job_runner
+
+    @classmethod
+    def encode(cls, data):
+        """ encode data for SocketConnection. """
+        return json.dumps(data).encode("utf-8")
+
+    @classmethod
+    def decode(cls, data):
+        """ decode data from SocketConnection. """
+        return json.loads(data.decode("utf-8"))
 
     @property
     def job_queue(self):
@@ -158,25 +201,25 @@ class SocketProtocol(object):
     def call(cls, connection, command, args):
         """ Send command, decode and return response. """
         raw_command = cls.commands.build(command, args)
-        connection.send(json.dumps(raw_command))
-        return json.loads(connection.recv())
+        connection.send(cls.encode(raw_command))
+        return cls.decode(connection.recv())
 
     def handle(self):
         data = self.connection.recv()
         try:
-            command, args = self.commands.parse(json.loads(data))
+            command, args = self.commands.parse(self.decode(data))
         except Exception as e:
             logger.warn("Unknown command", exc_info=True)
-            self.respond("Unknown command: %s" % e)
+            self.respond("Unknown command: {}".format(e))
         try:
             logger.info("Running %s", command.name)
             command(self, *args)
         except Exception as e:
-            logger.error("Command failed: %r", command, exc_info=True)
-            self.respond("Error: %s" % e)
+            logger.error("Command failed: %s", repr(command), exc_info=True)
+            self.respond("Error: {}".format(e))
 
     def respond(self, data):
-        self.connection.send(json.dumps(data))
+        self.connection.send(self.encode(data))
 
     @commands.add('RELOAD')
     def __reload(self):
@@ -328,6 +371,7 @@ class SocketServer(object):
         if not self.socket_path:
             raise RuntimeError("No socket_path set")
         self.socket = socket.socket(socket.AF_UNIX)
+        logger.info("socket-server binding to: %s", repr(self.socket_path))
         self.socket.bind(self.socket_path)
         self.socket.listen(1)
         self._is_listening = True
@@ -339,30 +383,31 @@ class SocketServer(object):
                 time.sleep(1)
                 continue
 
-            with closing(conn):
+            with contextlib.closing(conn):
                 context = SocketProtocol(SocketConnection(conn), job_runner)
                 context.handle()
 
     def ping_server(self):
+        """ Check if a server is already listening on this socket_path.  """
+        if not os.path.exists(self.socket_path):
+            return False
         try:
-            os.stat(self.socket_path)
-            if self.send_cmd("PING") == 'PONG':
+            if self.send_cmd("PING", jr_socket=self.socket_path) == 'PONG':
                 return True
-        except socket.error:   # No server seems to be running
-            print("WARNING: Removing stale socket")
+        except socket.error:
+            # We have a socket file, but there's no server listening
+            logger.warning("Removing stale socket: %s", repr(self.socket_path))
             os.unlink(self.socket_path)
-            pass
-        except OSError:        # File didn't exist
-            pass
         return False
 
     def cleanup(self):
         if not self._is_listening:
             return
-        try:
-            os.unlink(self.socket_path)
-        except OSError:
-            pass
+        if os.path.exists(self.socket_path):
+            try:
+                os.unlink(self.socket_path)
+            except OSError:
+                pass
 
     def __del__(self):
         self.cleanup()
@@ -377,10 +422,9 @@ class SocketServer(object):
         args = args or []
         signal.signal(signal.SIGALRM, signal_timeout)
         signal.alarm(timeout)
+
         try:
-            with closing(socket.socket(socket.AF_UNIX)) as sock:
-                sock.connect(jr_socket)
-                return SocketProtocol.call(SocketConnection(sock), command,
-                                           args)
+            with jr_connection(path=jr_socket) as conn:
+                return SocketProtocol.call(conn, command, args)
         finally:
             signal.alarm(0)
